# ChallengeTwoApi/Models/ViewModels/UserViewModel.cs
namespace ChallengeTwoApi.Models.ViewModels
{
    public class UserViewModel
    {
        public int Id { get; set; }
        public string FirstName { get; set; }
        public string LastName { get; set; }
        public string Name { get; set; }
        public string? Email { get; set; }
        public List<byte[]>? Files { get; set; }
    }
}