# src/app/api/models/user-view-model.ts
/* tslint:disable */
/* eslint-disable */
export interface UserViewModel {
  email?: null | string;
  files?: null | Array<string>;
  firstName?: null | string;
  id?: number;
  lastName?: null | string;
  name?: null | string;
}